Update linuxprivchecker.sh
Updated [-] and [+].
Fixed Line 215 (now line 211 to 233).  The code is more readable.
Updated the line to be the same  ####.

If this looks good, I can work on the Python one to look about (if not) the same like this one.  This would be an opportunity for me to get better at Python.